==> Rscripts/03-modelling_species_richness/02-log_and_scaled_richness.R <==



# library -------------------------------------------------------------
  library(raster)
  library(tidyverse)

  rm(list = ls())

# data ----------------------------------------------------------------
  setwd()
  
  current.list <- list.files(path = "Results/rasters/iNEXT",
                             pattern = ".grd", full.names = T)
  names <- gsub(pattern = "Results/rasters/iNEXT/|.grd$", "", current.list)
  c.stack <- stack(current.list)
  names(c.stack) <- names
  list2env(setNames(unstack(c.stack), names(c.stack)), .GlobalEnv)

# relative richness -----------------------------------------------    
# requires: raster -- loaded from above
#           raster.name -- identity for file location, so maybe put in if its observed or predicted, and in quotes
  prop_sr <- function(raster, raster.name) {
  
    log_t <- log(raster)
    prop <- log_t/cellStats(log_t, stat = 'max', na.rm = T)
  
    rasterfile <- paste0("Results/rasters/scaled/", raster.name, ".grd")
  
  writeRaster(prop, rasterfile, overwrite = T)
  return(plot(prop))
  
  }
  
# native
  prop_sr(native_C3, "native_C3")
  prop_sr(native_C4, "native_C4")
  prop_sr(native_total, "native_total")
# non-native (exotic)
  prop_sr(nonnative_C3, "nonnative_C3")
  prop_sr(nonnative_C4, "nonnative_C4")
  prop_sr(nonnative_total, "nonnative_total")

 # -------------------------------------------------------------------
